requirements.txt
fastapi
uvicorn
openai
twilio
python-dotenv
